+++ utils/build_badges.py
@@ -47,7 +47,7 @@ def main():
     stats["coverage"] = f"{percent_covered}%"
     stats["pypi"] = pkg_version  # get_pkg_version_latest_pypi()
     stats["license"] = "BSD-3"
-    stats["code-style"] = "black"
+    stats["code-style"] = "ruff"
     stats["python"] = "3.9+"
 
     coverage_thresholds = {0.6: "red", 0.7: "orange", 0.9: "yellow", 1: "green"}